requirements.txt
numpy<2.0.0
haystack-ai 